requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==8.0.3
cycler==0.10.0
Cython==0.29.21
filelock==3.0.12
Flask==2.0.2
gast==0.3.3
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
picamera==1.13
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pyzbar==0.1.8
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow @ file:///tensorflow-2.2.0-cp37-cp37m-linux_armv7l.whl
tensorflow-estimator==2.2.0
termcolor==1.1.0
tqdm==4.54.0
urllib3==1.26.2
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.4.0
##pygame==1.9.4
